* added: new config/edit mode inside inventory/magic and favourite menu
the bound button press will add the selected item to the list
* added: new config for non elden souls mode as well
* adjusted: code internal cleanup
* fixed: using version 1.88 of imgui again, seems to work better
currently for the project
* cleanup: removed outdated mcm options and config
* added: persist if ui was hidden or shown with the button press
* added: settings to cleanup config if the player does not have
items/spells anymore, can be disabled/enabled for each supported type.
(Cleanup Settings>Clean)
* added: option to enter edit/config in menu with a key, should prevent
if the same keys are bound to the HUD that can be used as menu movement
as well (Controls>Optional Controls>Require Key To Enter Edit & Edit
Key)
* added: check and offset if the player has a shout
* added: replace handling for elden souls, it can only be replaced the
current active and shown setting (toggle key down + bound key press in
inventory/magic/favorite menu will if the selected item/spell is
suitable)
* fixed: do not call to consume a potion if there was no potion found
* changed: ini debug setting changed to bool
* added: ini setting to disable font loading (Font>bFontLoad)
* cleanup: internal code cleanup and adjustments
* added: rewrite of config in elden souls if auto cleanup is enabled
* added: check if the given font file is the correct file type
* cleanup: removed not called and needed functions
* added: inside the MCM "Information Bind" and "Information Config" some
simple Information "how to" and "what can be added"
* fixed: added a check if we have settings to execute after button press
* cleanup: separated key inputs from hud and menu config
* added: basic apply of poison to equipped weapon
* added: Controls for new Edit/config
* adjusted: Renamed some Translations
* adjusted: some controll inputs will be blocked if toggle/edit or edit
2 button are down in the inventory, magic and favorite menu. This should
make controller config better
* adjusted: removing poison now after usage
* adjusted: added a check to the location needs block check, no point in
doing that if the game is paused
* added: cleanup setting, if enabled, will check at new game/post game
load and if an inventory item that is configured and new count == 0
* adjusted: If Items are not cleaned, but the player currently does not
hold the item/spell it will be marked as "blocked" (= Graphic
Setting>Icon Blocked)
* adjusted: cleaned up translation string so they have the same naming
* added: "Remove Key" (Controls>Supporting Controls>Remove Key). In the
Inventory/Magic or Favorite Menu Button Down + Any of the four Action
Key will look for the form in ALL Settings if found, it will be removed.
* adjusted: instant spell casting
* fixed: added missing check for the remove button to block input for
skyui
* fixed: if an item gets removed, make sure the setting is overwritten
for the ui as well

+++ src/setting/custom_setting.cpp
@@ -2,7 +2,6 @@
 #include "file_setting.h"
 #include "mcm_setting.h"
 #include "util/constant.h"
-#include <SimpleIni.h>
 
 namespace config {
     CSimpleIniA custom_ini;
@@ -76,14 +75,14 @@ namespace config {
 
     void custom_setting::write_slot_action_by_section(const std::string& a_section, const uint32_t a_action) {
         read_setting();
-        custom_ini.SetLongValue(a_section.c_str(), "uSlotAction", a_action);
+        custom_ini.SetLongValue(a_section.c_str(), "uSlotAction", static_cast<long>(a_action));
 
         save_setting();
     }
 
     void custom_setting::write_slot_action_left_by_section(const std::string& a_section, const uint32_t a_action) {
         read_setting();
-        custom_ini.SetLongValue(a_section.c_str(), "uSlotActionLeft", a_action);
+        custom_ini.SetLongValue(a_section.c_str(), "uSlotActionLeft", static_cast<long>(a_action));
 
         save_setting();
     }
@@ -117,16 +116,16 @@ namespace config {
 
         reset_section(section);
 
-        custom_ini.SetLongValue(section, "uPage", a_page);
-        custom_ini.SetLongValue(section, "uPosition", a_position);
-        custom_ini.SetLongValue(section, "uType", a_type);
+        custom_ini.SetLongValue(section, "uPage", static_cast<long>(a_page));
+        custom_ini.SetLongValue(section, "uPosition", static_cast<long>(a_position));
+        custom_ini.SetLongValue(section, "uType", static_cast<long>(a_type));
         custom_ini.SetValue(section, "sSelectedItemForm", a_form.c_str());
-        custom_ini.SetLongValue(section, "uSlotAction", a_action);
-        custom_ini.SetLongValue(section, "uHandSelection", a_hand);
+        custom_ini.SetLongValue(section, "uSlotAction", static_cast<long>(a_action));
+        custom_ini.SetLongValue(section, "uHandSelection", static_cast<long>(a_hand));
         custom_ini.SetLongValue(section, "iEffectActorValue", a_effect_actor_value);
-        custom_ini.SetLongValue(section, "uTypeLeft", a_type_left);
+        custom_ini.SetLongValue(section, "uTypeLeft", static_cast<long>(a_type_left));
         custom_ini.SetValue(section, "sSelectedItemFormLeft", a_form_left.c_str());
-        custom_ini.SetLongValue(section, "uSlotActionLeft", a_action_left);
+        custom_ini.SetLongValue(section, "uSlotActionLeft", static_cast<long>(a_action_left));
 
         save_setting();
     }
